refactor(PaymentRestClient): Http 헤더 의존성 주입 형식으로 변경
